1. use pyaudio instead of pygame to play wave files 2. yapper/speaker.py -> PiperSpeaker, use highest available quality of the given voice by default

+++ yapper/speaker.py
@@ -1,7 +1,10 @@
 import os
 import subprocess
+import wave
 from abc import ABC, abstractmethod
+from typing import Optional
 
+import pyaudio
 import pyttsx3 as tts
 
 import yapper.constants as c
@@ -9,14 +12,26 @@ from yapper.enums import PiperQuality, PiperVoice
 from yapper.utils import (
     APP_DIR,
     PLATFORM,
-    install_piper,
     download_piper_model,
-    get_random_name
+    get_random_name,
+    install_piper
 )
 
-# suppresses pygame's welcome message
-os.environ["PYGAME_HIDE_SUPPORT_PROMPT"] = "1"
-import pygame  # noqa: E402
+
+def play_wave(pa_instance: pyaudio.PyAudio, wave_f: str):
+    with wave.open(wave_f, "rb") as wf:
+        stream = pa_instance.open(
+            format=pa_instance.get_format_from_width(wf.getsampwidth()),
+            channels=wf.getnchannels(),
+            rate=wf.getframerate(),
+            output=True,
+        )
+        data = wf.readframes(1024)
+        while data:
+            stream.write(data)
+            data = wf.readframes(1024)
+        stream.stop_stream()
+        stream.close()
 
 
 class BaseSpeaker(ABC):
@@ -75,26 +90,49 @@ class DefaultSpeaker(BaseSpeaker):
 
 
 class PiperSpeaker(BaseSpeaker):
+    VOICE_QUALITY_MAP = {
+        PiperVoice.AMY: PiperQuality.MEDIUM,
+        PiperVoice.ARCTIC: PiperQuality.MEDIUM,
+        PiperVoice.BRYCE: PiperQuality.MEDIUM,
+        PiperVoice.DANNY: PiperQuality.LOW,
+        PiperVoice.HFC_FEMALE: PiperQuality.MEDIUM,
+        PiperVoice.HFC_MALE: PiperQuality.MEDIUM,
+        PiperVoice.JOE: PiperQuality.MEDIUM,
+        PiperVoice.JOHN: PiperQuality.MEDIUM,
+        PiperVoice.KATHLEEN: PiperQuality.LOW,
+        PiperVoice.KRISTIN: PiperQuality.MEDIUM,
+        PiperVoice.KUSAL: PiperQuality.MEDIUM,
+        PiperVoice.L2ARCTIC: PiperQuality.MEDIUM,
+        PiperVoice.LESSAC: PiperQuality.HIGH,
+        PiperVoice.LIBRITTS: PiperQuality.HIGH,
+        PiperVoice.LIBRITTS_R: PiperQuality.MEDIUM,
+        PiperVoice.LJSPEECH: PiperQuality.HIGH,
+        PiperVoice.NORMAN: PiperQuality.MEDIUM,
+        PiperVoice.RYAN: PiperQuality.HIGH,
+    }
+
     def __init__(
         self,
         voice: PiperVoice = PiperVoice.AMY,
-        quality: PiperQuality = PiperQuality.MEDIUM,
+        quality: Optional[PiperQuality] = None,
     ):
         """
         Speaks the text using piper.
 
         Parameters
         ----------
-        voice : str, optional
+        voice : PiperVoice, optional
             Name of the piper voice to be used, can be one of 'PiperVoice'
             enum's attributes (default: PiperVoice.AMY).
-        quality : str, optional
+        quality : PiperQuality, optional
             Quality of the voice, can be ont of 'PiperQuality'
-            enum's attributes (default: PiperQuality.MEDIUM).
+            enum's attributes (default: the highest available quality of
+            the given voice).
         """
         assert (
             voice in PiperVoice
         ), f"voice must be one of {', '.join(PiperVoice)}"
+        quality = quality or PiperSpeaker.VOICE_QUALITY_MAP[voice]
         assert (
             quality in PiperQuality
         ), f"quality must be one of {', '.join(PiperQuality)}"
@@ -108,11 +146,10 @@ class PiperSpeaker(BaseSpeaker):
             voice.value, quality.value
         )
         self.onnx_f, self.conf_f = str(self.onnx_f), str(self.conf_f)
-        pygame.mixer.init()
+        self.pa_instance = pyaudio.PyAudio()
 
-    def say(self, text: str):
-        """Speaks the given text"""
-        f = APP_DIR / f"{get_random_name()}.wav"
+    def text_to_wave(self, text: str, file: str):
+        """Saves the speech for the given text into the given file."""
         subprocess.run(
             [
                 self.exe_path,
@@ -121,15 +158,20 @@ class PiperSpeaker(BaseSpeaker):
                 "-c",
                 self.conf_f,
                 "-f",
-                str(f),
+                file,
                 "-q",
             ],
             input=text.encode("utf-8"),
-            check=True,
             stdout=subprocess.DEVNULL,
+            check=True,
         )
-        sound = pygame.mixer.Sound(f)
-        sound.play()
-        while pygame.mixer.get_busy():
-            pygame.time.wait(100)
-        os.remove(f)
+
+    def say(self, text: str):
+        """Speaks the given text"""
+        f = APP_DIR / f"{get_random_name()}.wav"
+        try:
+            self.text_to_wave(text, str(f))
+            play_wave(self.pa_instance, str(f))
+        finally:
+            if f.exists():
+                os.remove(f)
